chore(Auditable): rename stub class
